small code better song

+++ OperationLIbrary/IOperazione.cs
@@ -8,12 +8,12 @@ namespace OperationLIbrary
 {
     public interface IOperazione
     {
-        int Somma(int n);
-        int Sottrazione(int n);
-        int Divisione(int n);
-        int Moltiplicazione(int n);
-        int XAllaseconda(int n);
-        double Radicequadrata(int n);
-        double Unosux(int n);
+        double Somma(double n);
+        double Sottrazione(double n);
+        double Divisione(double n);
+        double Moltiplicazione(double n);
+        double XAllaseconda(double n);
+        double Radicequadrata(double n);
+        double Unosux(double n);
     }
 }
